python/requirements.txt
certifi==2024.2.2
influxdb-client==1.40.0
numpy==1.26.4
pandas==2.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
reactivex==4.0.4
setuptools==69.0.3
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wheel==0.42.0
